feat: Remove Passing Explicit Arguments
Rather than explicitly destructuring arguments and passing them individually to the send function of a given notification middleware, the entire arguments object, execpt `config` and `modifiers` (which anyway is passed specific to the notification protocol) properties, is passed instead. This way, the middleware is immune to the changes in arguments in any specific protocol middleware when sending notifications.